Generic item menu
Refato menu inventory pour être utilisé par RemovalCurse, Identified, et autres items use on item.
